reorganized packages, configured to allow autowiring from different packages
