+++ src/main/java/com/example/demo/R4jDemoFrontApplication.java
@@ -2,9 +2,10 @@ package com.example.demo;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
-
+@ComponentScan(basePackages = {"com.example.demo", "com.example.properties"})
 public class R4jDemoFrontApplication {
 
 	public static void main(String[] args) {
