+++ src/main/java/com/example/properties/CircuitBreakerCustomConfig.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.properties;
 
 import java.time.Duration;
