+++ src/main/java/com/example/properties/RetryCustomConfig.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.properties;
 
 import java.time.Duration;
